Use printf("%zu") to print size_t

+++ src/rar2john.c
@@ -82,9 +82,9 @@ static int process_file5(const char *archive_name);
 static int check_fread(const size_t buf_size, const size_t size, const size_t nmemb)
 {
 	if (buf_size < size * nmemb) {
-		fprintf(stderr, "Error: check_fread(buf_size=%lu, size=%lu, nmemb=%lu) failed, "
+		fprintf(stderr, "Error: check_fread(buf_size=%zu, size=%zu, nmemb=%zu) failed, "
 			"buf_size is smaller than size * nmemb.\n",
-			(unsigned long)buf_size, (unsigned long)size, (unsigned long)nmemb);
+			buf_size, size, nmemb);
 		return 0;
 	}
 	return 1;
